Extract SanitizingFormatter to reduce code complexity
